finished: Don't try to blink keyboard LEDs
Previously, in unattended mode, on error, the keyboard LEDs would blink
every second, in hope of attracting an operator's attention. Neat!

However, this depended on X11-specific API, and did not check that the
app is running as an X client rather than a Wayland client. As a result,
on Wayland systems, the app would segfault 1 second after displaying
an error in unattended mode, the first time it tried to adjust the LED
state.

Remove this behaviour.