Adjust for loopback-based image boot path
We have simplified our image boot path, eliminating use of device-mapper.
/dev/disk/endless-image has been specifically created to enable the
installer to find the source image data, equivalent to the previous
/dev/mapper/endless-image node.

Without device mapper getting in the way, we can now directly mount the
"image device" partition that holds endless.img in all cases. However,
the endless-image link is still needed for the squashfs case,
which eos-installer does not know how to handle itself.

There's some other code flow and comments in this file that could be
updated for this new scheme, but this is the minimal increment to keep
things working.

https://phabricator.endlessm.com/T35126