Mount iso9660 partitions read-only
On live boots where we now boot directly from loopback (without having
an additional device-mapper layer), we need to mount (for a 2nd time)
the image partition in order to find the data needed by the installer.

For some reason, the mount command fails to handle this case, because
it (per its default) tries a rw mount and gets EBUSY from the kernel,
then it gives up.

If you try to double-mount an iso9660 partition in a simple test you
get different behaviour: for both mounts, it attempts rw first as normal,
the kernel returns EACCES and mount falls back to 'ro'.

Not exactly sure why the behaviour is different in this context,
however forcing read-only mounts for iso9660 does avoid the issue.

https://phabricator.endlessm.com/T35126