driver: Remove small-screen support
This code, inherited from GNOME Initial Setup, attempts to maximise the
window & add a scrollbar when needed if the screen is very small, which
is defined to be less than 600 pixels tall or 800 pixels wide.

By today's standards this is incredibly small. The code here also
doesn't try to react to changes in the screen resolution. And the window
itself is just not very big.

Remove it. This has the side benefit of removing calls to
gdk_screen_height() and gdk_screen_width() which are deprecated.