eosimages.mount: Remove allow_other flag
On standalone eosinstaller images, the eosimages partition is mounted at
a fixed path by systemd. This is (among other things) for the benefit of
Endless OS's patched GNOME Initial Setup checking for an unattended.ini
file at a fixed path, in order to skip the language selector if it is
present.

Historically, exfat was supported via FUSE. FUSE filesystems are, by
default, only accessible to the user that is running the userspace
filesystem process, regardless of the Unix permissions on the file. And
since this filesystem was mounted as root, the gnome-initial-setup user
(which is what runs eos-installer on eosinstaller images, as well as
GNOME Initial Setup of course) would not be able to feast on the files
within.

However, we now use the in-kernel exfat implementation. As a result,
this mount unit always failed. This didn't break eos-installer on
eosinstaller images because it knows how to use UDisks to enumerate
available filesystems and mount the most likely-looking one, but it did
mean that Endless OS's patched GNOME Initial Setup could not locate the
unattended.ini file, meaning that the install process became attended:
you had to click Next on the language selector page, while the intention
is that no action is necessary beyond booting from the USB drive.

https://phabricator.endlessm.com/T35620